fix(phase-3): resolve mypy-strict friction in lenia (Taichi override + sim.py)
Closes the 13 mypy-strict errors that have red-lit the `test-lenia`
job's `mypy --strict (lenia)` step on every push since `5baf083`
(lenia Stage 1b feat-push). At HEAD `33359fc` the failure was the
only red step in the entire CI surface (10 workflows green save this
one job's mypy step — selective-LFS-fetch was newly green after
r2-credentials-durability + audit-citation-hygiene chains).

THREE distinct mypy errors, three independent fixes:

**(a) 11 errors in `lenia/_taichi_kernels.py`** — `@ti.kernel` is an
untyped decorator (`[untyped-decorator]`), Taichi propagates `Any` so
the kernel function becomes `[no-untyped-def]`, and per-arg
`ti.types.ndarray(dtype=ti.f64, ndim=2)` annotations are function
CALLS (not types) so `[valid-type]` fires on each. All three error
classes are STRUCTURAL to Taichi's AST-transformer kernel definition
(IC-12 § 4.2 / § 4.6) — PEP-563 stringification would break
`@ti.kernel` at decoration time. There is NO type-friendlier syntax
Taichi supports; the precedent file
`packages/reaction-diffusion-2d-stack-d/reaction_diffusion_2d_stack_d/reference/gray_scott_taichi.py`
has the IDENTICAL shape but is NOT in any CI workflow's mypy scope,
so this is the **first CI-strict-mypy-vs-Taichi collision** in the
repo and there is no inheritable precedent to copy verbatim.

Fix: per-module override at `packages/lenia/pyproject.toml`
`[[tool.mypy.overrides]]` block targeting `lenia._taichi_kernels` with
`disable_error_code = ["misc", "no-untyped-def", "untyped-decorator",
"valid-type"]`. Narrow over inline `# type: ignore` — keeps the
kernel module declaration verbatim and confines the suppression to
the four affected codes (rest of strict mode active). The rest of
`lenia/` remains under full `strict = true`. Sets the inheritable
precedent for future Phase-3-D Taichi-kernel CI scoping.

**(b) `lenia/sim.py:82` `[return-value]`** — `(K / total)` returns
`ndarray[..., floating[Any]]` per NumPy 2.x stubs (scalar promotion
widens the dtype despite K being `float64` and `total` being
`float`). Function signature is `NDArray[np.float64]`.

Fix: `(K / total).astype(np.float64, copy=False)` — explicit dtype
preservation with zero-copy when dtype matches at runtime. Honest
signature; mathematically correct; no `cast()` lie.

**(c) `lenia/sim.py:134` + `:182` `[unused-ignore]`** — two
`# type: ignore[import-not-found]` comments on `from common_py.*`
imports. The pyproject's `[[tool.mypy.overrides]] module = [...,
"common_py", "common_py.*", ...] ignore_missing_imports = true`
block already silences these; the inline ignores are vestigial.

Fix: delete the two ignore-comments. The imports resolve correctly
via the workspace.

Verification (local Python 3.12.3 + uv-synced workspace):
`uv run mypy --strict lenia/` — `Success: no issues found in 6
source files`. `ruff check lenia/ tests/` + `ruff format --check
lenia/ tests/` clean. Pytest sees a PRE-EXISTING DeprecationWarning
on `locale.getdefaultlocale()` in this local Python 3.12.3 patch
(also fires at HEAD `33359fc` PRIOR to this fix per `git stash`
spot-check) — not introduced by this commit, surfaces independently
on the post-fix CI poll under Python 3.12.13. Out-of-scope for this
fix; if CI surfaces the same warning, separate dispatch.

Integrity invariant `0 HARD_FAIL / 14 SOFT_WARN` HELD per §R; live
digest at HEAD `33359fc` = `688bc195d8b785753ae9500b4e1d48800ae961dd38ac4410f16fb7446de127ff`.

+++ packages/lenia/lenia/sim.py
@@ -79,7 +79,11 @@ def _build_kernel_window(R: int) -> NDArray[np.float64]:
     total = float(K.sum())
     if total <= 0.0:
         raise ValueError(f"Quad4 kernel window sum is non-positive ({total})")
-    return K / total
+    # ``K`` is float64; dividing by a Python ``float`` is mathematically
+    # dtype-preserving, but NumPy 2.x stubs widen the result to
+    # ``ndarray[..., floating[Any]]``. Pin the dtype explicitly (no-op copy
+    # since the underlying dtype matches) so the signature stays honest.
+    return (K / total).astype(np.float64, copy=False)
 
 
 def _init_orbium_field(grid: int, seed: int) -> NDArray[np.float64]:
@@ -131,7 +135,7 @@ class LeniaSim:
     def _ensure_taichi(self) -> None:
         if self._taichi_initialized:
             return
-        from common_py.determinism import (  # type: ignore[import-not-found]
+        from common_py.determinism import (
             Config as DeterminismConfig,
         )
         from common_py.determinism import set_taichi_deterministic
@@ -179,7 +183,7 @@ class LeniaSim:
         Consumes :class:`common_py.capture.Writer` with the IC-2 API
         (``write_step(idx, data)`` + ``finalize()``).
         """
-        from common_py.capture import (  # type: ignore[import-not-found]
+        from common_py.capture import (
             ConfigMeta,
             DeterminismMeta,
             Manifest,
